Made the Navbar for the Website
I made the navbar for the website. Instead of using AirBNB's logo, I
used an image that I found off of the internet for the letter A (which
stands for my first name). I also styled the logo to fit into the
navbar. Finally, I styled the navbar to have a box-shadow.