Finished Hero Section of the Website
I finished making the hero section of the website. I added an image
of a photo collage for aesthetic purposes. Next, I added a heading text
and paragraph text. Finally, I styled the added components as needed.

P.S: I also started to make the card section of the website.