Finishing Touches for Cards
I added the finishing touches for the cards at the bottom of the page.
I centered the cards, added a badge that changes depending on whether
the event is sold out, online, or neither (For example, if the event
was an online event that was sold out, it would just say sold out), and
I made minor changes to the card stats (for example, the ratings were
changed).

My replica of the 'AirBNB Experiences' page is complete.